Enable shared library build.
* This patch can enable Polygeist as shared library. aka, when cmake
    config append with `-DBUILD_SHARED_LIBS=ON`.

* Change is simple, just add missing libray and move two functions
    to Ops.cpp from OpenMPOpt.cpp.
